+++ packages/frontend/src/__mocks__/api/index.js
@@ -1,0 +1,4 @@
+import { setupServer } from 'msw/node';
+import { handlers } from './handlers';
+
+export const mswServer = setupServer(...handlers);
