+++ packages/frontend/src/__mocks__/api/mockPhotos.js
@@ -1,0 +1,147 @@
+export const mockPhotos = [
+  {
+    id: 1687678,
+    width: 2832,
+    height: 2832,
+    url: 'https://www.pexels.com/photo/close-up-photo-of-small-jellyfish-1687678/',
+    photographer: 'Valdemaras D.',
+    photographer_url: 'https://www.pexels.com/@valdemaras-d-784301',
+    photographer_id: 784301,
+    avg_color: '#222C2F',
+    src: {
+      original:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg',
+      large2x:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
+      large:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
+      medium:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&h=350',
+      small:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&h=130',
+      portrait:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
+      landscape:
+        'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
+      tiny: 'https://images.pexels.com/photos/1687678/pexels-photo-1687678.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
+    },
+    liked: false,
+    alt: 'Close-Up Photo of Small Jellyfish',
+  },
+  {
+    id: 3493777,
+    width: 3000,
+    height: 3000,
+    url: 'https://www.pexels.com/photo/photo-of-bridge-under-white-clouds-3493777/',
+    photographer: 'Aleksey Kuprikov',
+    photographer_url: 'https://www.pexels.com/@aleksey-kuprikov-1883853',
+    photographer_id: 1883853,
+    avg_color: '#6B7874',
+    src: {
+      original:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg',
+      large2x:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
+      large:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
+      medium:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&h=350',
+      small:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&h=130',
+      portrait:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
+      landscape:
+        'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
+      tiny: 'https://images.pexels.com/photos/3493777/pexels-photo-3493777.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
+    },
+    liked: false,
+    alt: 'Photo of Bridge Under White Clouds',
+  },
+  {
+    id: 1683492,
+    width: 3648,
+    height: 3648,
+    url: 'https://www.pexels.com/photo/ocean-wave-splashing-on-dock-1683492/',
+    photographer: 'Ricky Esquivel',
+    photographer_url: 'https://www.pexels.com/@rickyrecap',
+    photographer_id: 722822,
+    avg_color: '#8298A9',
+    src: {
+      original:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg',
+      large2x:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
+      large:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
+      medium:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&h=350',
+      small:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&h=130',
+      portrait:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
+      landscape:
+        'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
+      tiny: 'https://images.pexels.com/photos/1683492/pexels-photo-1683492.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
+    },
+    liked: false,
+    alt: 'Ocean Wave Splashing on Dock',
+  },
+  {
+    id: 789555,
+    width: 3327,
+    height: 3327,
+    url: 'https://www.pexels.com/photo/woman-in-black-hijab-headscarf-walking-on-field-789555/',
+    photographer: 'Admiral General M. ⭐⭐⭐⭐⭐ GodShepherdly 33277089*',
+    photographer_url: 'https://www.pexels.com/@samsilitongajr',
+    photographer_id: 218404,
+    avg_color: '#A19584',
+    src: {
+      original:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg',
+      large2x:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
+      large:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
+      medium:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&h=350',
+      small:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&h=130',
+      portrait:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
+      landscape:
+        'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
+      tiny: 'https://images.pexels.com/photos/789555/pexels-photo-789555.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
+    },
+    liked: false,
+    alt: 'Woman in Black Hijab Headscarf Walking on Field',
+  },
+  {
+    id: 2894944,
+    width: 2359,
+    height: 2359,
+    url: 'https://www.pexels.com/photo/man-standing-on-grass-field-2894944/',
+    photographer: 'Gagaz Adam',
+    photographer_url: 'https://www.pexels.com/@ggzadam',
+    photographer_id: 398931,
+    avg_color: '#4A5B61',
+    src: {
+      original:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg',
+      large2x:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&dpr=2&h=650&w=940',
+      large:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&h=650&w=940',
+      medium:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&h=350',
+      small:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&h=130',
+      portrait:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=1200&w=800',
+      landscape:
+        'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&fit=crop&h=627&w=1200',
+      tiny: 'https://images.pexels.com/photos/2894944/pexels-photo-2894944.jpeg?auto=compress&cs=tinysrgb&dpr=1&fit=crop&h=200&w=280',
+    },
+    liked: false,
+    alt: 'Man Standing on Grass Field',
+  },
+];
